requirements.txt
Adafruit-Blinka==6.13.0
adafruit-circuitpython-74hc595==1.3.0
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-charlcd==3.3.10
adafruit-circuitpython-mcp230xx==2.5.0
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
aniso8601==9.0.1
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.8.24
cffi==1.14.6
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fourletterphat==0.1.0
gpiozero==1.6.2
html5lib==1.0.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
logilab-common==1.4.2
lxml==4.3.2
MarkupSafe==1.1.0
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
numpy==1.21.2
oauthlib==2.1.0
olefile==0.46
pandas==1.3.3
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==5.4.1
pkg-resources==0.0.0
psutil==5.5.1
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pyftdi==0.53.3
pygame==1.9.6
Pygments==2.3.1
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.2.2
pyOpenSSL==19.0.0
pyserial==3.4
python-dateutil==2.8.2
pytz==2021.1
pyusb==1.2.1
pyxdg==0.25
rainbowhat==0.1.0
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.5
ssh-import-id==5.7
sysv-ipc==1.1.0
thonny==3.3.10
touchphat==0.0.1
twilio==6.63.2
twython==3.7.0
typed-ast==1.3.1
unicornhathd==0.0.4
urllib3==1.24.1
w1thermsensor==2.0.0
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
